+++ runner/src/mill/runner/MillServerMain.scala
@@ -7,33 +7,13 @@ import java.net.Socket
 import scala.jdk.CollectionConverters._
 import org.newsclub.net.unix.AFUNIXServerSocket
 import org.newsclub.net.unix.AFUNIXSocketAddress
-import mill.main.BuildInfo
 import mill.main.client._
 import mill.api.{SystemStreams, internal}
 import mill.main.client.ProxyStream.Output
 import mill.main.client.lock.{Lock, Locks}
-
 import scala.util.Try
 
-@internal
-trait MillServerMain[T] {
-  def stateCache0: T
-  var stateCache = stateCache0
-  def main0(
-      args: Array[String],
-      stateCache: T,
-      mainInteractive: Boolean,
-      streams: SystemStreams,
-      env: Map[String, String],
-      setIdle: Boolean => Unit,
-      systemProperties: Map[String, String],
-      initialSystemProperties: Map[String, String]
-  ): (Boolean, T)
-}
-
-@internal
-object MillServerMain extends MillServerMain[RunnerState] {
-  def stateCache0 = RunnerState.empty
+object MillServerMain{
   def main(args0: Array[String]): Unit = {
     // Disable SIGINT interrupt signal in the Mill server.
     //
@@ -52,101 +32,46 @@ object MillServerMain extends MillServerMain[RunnerState] {
     val acceptTimeoutMillis =
       Try(System.getProperty("mill.server_timeout").toInt).getOrElse(5 * 60 * 1000) // 5 minutes
 
-    new Server(
+    new MillServerMain(
       lockBase = os.Path(args0(0)),
-      this,
       () => System.exit(Util.ExitServerCodeWhenIdle()),
       acceptTimeoutMillis = acceptTimeoutMillis,
       Locks.files(args0(0))
     ).run()
   }
+}
+class MillServerMain(
+                      lockBase: os.Path,
+                      interruptServer: () => Unit,
+                      acceptTimeoutMillis: Int,
+                      locks: Locks
+                    )
+  extends mill.main.server.MillServerMain[RunnerState](lockBase, interruptServer, acceptTimeoutMillis, locks) {
+  def stateCache0 = RunnerState.empty
+
 
   def main0(
-      args: Array[String],
-      stateCache: RunnerState,
-      mainInteractive: Boolean,
-      streams: SystemStreams,
-      env: Map[String, String],
-      setIdle: Boolean => Unit,
-      userSpecifiedProperties: Map[String, String],
-      initialSystemProperties: Map[String, String]
-  ): (Boolean, RunnerState) = {
+             args: Array[String],
+             stateCache: RunnerState,
+             mainInteractive: Boolean,
+             streams: SystemStreams,
+             env: Map[String, String],
+             setIdle: Boolean => Unit,
+             userSpecifiedProperties: Map[String, String],
+             initialSystemProperties: Map[String, String]
+           ): (Boolean, RunnerState) = {
     try MillMain.main0(
-        args = args,
-        stateCache = stateCache,
-        mainInteractive = mainInteractive,
-        streams0 = streams,
-        bspLog = None,
-        env = env,
-        setIdle = setIdle,
-        userSpecifiedProperties0 = userSpecifiedProperties,
-        initialSystemProperties = initialSystemProperties
-      )
+      args = args,
+      stateCache = stateCache,
+      mainInteractive = mainInteractive,
+      streams0 = streams,
+      bspLog = None,
+      env = env,
+      setIdle = setIdle,
+      userSpecifiedProperties0 = userSpecifiedProperties,
+      initialSystemProperties = initialSystemProperties
+    )
     catch MillMain.handleMillException(streams.err, stateCache)
-  }
-}
-
-class Server[T](
-    lockBase: os.Path,
-    sm: MillServerMain[T],
-    interruptServer: () => Unit,
-    acceptTimeoutMillis: Int,
-    locks: Locks
-) {
-
-  val originalStdout = System.out
-  val serverId = scala.util.Random.nextLong().toString
-  def serverLog(s: String) = os.write.append(lockBase / ServerFiles.serverLog, s + "\n")
-  def run(): Unit = {
-
-    val initialSystemProperties = sys.props.toMap
-    Server.tryLockBlock(locks.processLock) {
-      watchServerIdFile()
-
-      var running = true
-      while (running) {
-
-        val socketPath = os.Path(ServerFiles.pipe(lockBase.toString()))
-
-        os.remove.all(socketPath)
-
-        // Use relative path because otherwise the full path might be too long for the socket API
-        val addr =
-          AFUNIXSocketAddress.of(socketPath.relativeTo(os.pwd).toNIO.toFile)
-        serverLog("listening on socket")
-        val serverSocket = AFUNIXServerSocket.bindOn(addr)
-        val socketClose = () => serverSocket.close()
-
-        val sockOpt = interruptWith(
-          "MillSocketTimeoutInterruptThread",
-          acceptTimeoutMillis,
-          socketClose(),
-          serverSocket.accept()
-        )
-
-        sockOpt match {
-          case None => running = false
-          case Some(sock) =>
-            try {
-              serverLog("handling run")
-              try handleRun(sock, initialSystemProperties)
-              catch { case e: Throwable => serverLog(e + "\n" + e.getStackTrace.mkString("\n"))}
-              finally sock.close();
-            } finally serverSocket.close()
-        }
-      }
-    }.getOrElse(throw new Exception("PID already present"))
-  }
-
-  def proxyInputStreamThroughPumper(in: InputStream): PipedInputStream = {
-    val pipedInput = new PipedInputStream()
-    val pipedOutput = new PipedOutputStream()
-    pipedOutput.connect(pipedInput)
-    val pumper = new InputPumper(() => in, () => pipedOutput, false)
-    val pumperThread = new Thread(pumper, "proxyInputStreamThroughPumper")
-    pumperThread.setDaemon(true)
-    pumperThread.start()
-    pipedInput
   }
 
   def handleRun(clientSocket: Socket, initialSystemProperties: Map[String, String]): Unit = {
@@ -186,9 +111,9 @@ class Server[T](
       val t = new Thread(
         () =>
           try {
-            val (result, newStateCache) = sm.main0(
+            val (result, newStateCache) = main0(
               args,
-              sm.stateCache,
+              stateCache,
               interactive,
               new SystemStreams(stdout, stderr, proxiedSocketInput),
               env.asScala.toMap,
@@ -197,7 +122,7 @@ class Server[T](
               initialSystemProperties
             )
 
-            sm.stateCache = newStateCache
+            stateCache = newStateCache
             os.write.over(
               lockBase / ServerFiles.exitCode,
               (if (result) 0 else 1).toString.getBytes()
@@ -233,78 +158,5 @@ class Server[T](
 
     } finally ProxyStream.sendEnd(currentOutErr) // Send a termination
   }
-
-  def watchServerIdFile() = {
-    os.write.over(lockBase / ServerFiles.serverId, serverId)
-    val serverIdThread = new Thread(
-      () => {
-        while (true){
-          Thread.sleep(100)
-          Try(os.read(lockBase / ServerFiles.serverId)).toOption match{
-            case None =>
-              serverLog("serverId file missing, exiting")
-              System.exit(0)
-            case Some(s) =>
-              if (s != serverId){
-                serverLog(s"serverId file contents $s does not match serverId $serverId, exiting")
-                System.exit(0)
-              }
-          }
-        }
-      },
-      "Server ID Checker Thread"
-    )
-    serverIdThread.start()
-  }
-
-  def interruptWith[T](threadName: String, millis: Int, close: => Unit, t: => T): Option[T] = {
-    @volatile var interrupt = true
-    @volatile var interrupted = false
-    val thread = new Thread(
-      () => {
-        try Thread.sleep(millis)
-        catch { case t: InterruptedException => /* Do Nothing */ }
-        if (interrupt) {
-          interrupted = true
-          serverLog(s"Interrupting after ${millis}ms")
-          close
-        }
-      },
-      threadName
-    )
-
-    thread.start()
-    try {
-      val res =
-        try Some(t)
-        catch { case e: Throwable => None }
-
-      if (interrupted) None
-      else res
-
-    } finally {
-      thread.interrupt()
-      interrupt = false
-    }
-  }
-}
-
-object Server {
-
-  def lockBlock[T](lock: Lock)(t: => T): T = {
-    val l = lock.lock()
-    try t
-    finally l.release()
-  }
-
-  def tryLockBlock[T](lock: Lock)(t: => T): Option[T] = {
-    lock.tryLock() match {
-      case null => None
-      case l =>
-        try Some(t)
-        finally l.release()
-    }
-  }
-
 
 }
